switch ExtractCreationInfo's height to Int; switch Box.creationHeight to Int;

+++ src/main/scala/sigmastate/eval/CostingDataContext.scala
@@ -77,11 +77,11 @@ class CostingBox(val IR: Evaluation,
     } else
       super.getReg(i)
 
-  override def creationInfo: (Long, Col[Byte]) = {
+  override def creationInfo: (Int, Col[Byte]) = {
     import Types._
-    this.R3[(Long, Col[Byte])].get.asInstanceOf[Any] match {
+    this.R3[(Int, Col[Byte])].get.asInstanceOf[Any] match {
       case ConstantNode(arr: Array[Any], STuple(IndexedSeq(SLong, SByteArray))) if arr.length == 2 =>
-        (arr(0).asInstanceOf[Long], builder.Cols.fromArray(arr(1).asInstanceOf[Array[Byte]]))
+        (arr(0).asInstanceOf[Int], builder.Cols.fromArray(arr(1).asInstanceOf[Array[Byte]]))
       case v =>
         sys.error(s"Invalid value $v of creationInfo register R3")
     }
